requirements.txt
tqdm
pandas
plotly
openai
faiss-cpu
streamlit
python-dotenv
sentence-transformers